+++ activestorage/app/models/active_storage/variant.rb
@@ -103,10 +103,12 @@ class ActiveStorage::Variant
     service.delete(key)
   end
 
+  # Returns true if the variant has already been processed and uploaded to the service.
+  def processed?
+    service.exist?(key)
+  end
+
   private
-    def processed?
-      service.exist?(key)
-    end
 
     def process
       blob.open do |input|